Convert AuthContext to TypeScript with proper type definitions
